+++ docs/CHANGELOG.md
@@ -1,8 +1,11 @@
 # Webby Changelog
 
 # [Unreleased]
+## Fixed
+- history could not be opened
 
 ## [0.1.0] / alpha - 2020-09-15
+## Added
 - init release
 [Unreleased]:   https://github.com/lockieluke/webby/compare/alpha...HEAD
 [0.1.0]:        https://github.com/lockieluke/webby/releases/tag/alpha